chore(backend): move prisma.js to db directory
